Refactor: Move common bases setup code to CommonSetup
